engine/pyproject.toml
[tool.poetry]
name = "brick-engine"
version = "0.1.0"
description = "A better Elite Dangerous experience"
authors = ["Igor Ovsyannikov <kamish@outlook.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "brick_engine" }
]

[tool.poetry.dependencies]
python = "<3.13,>=3.8"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-socketio = "^5.11.2"
aiofiles = "^23.2.1"
pyinstaller = "^6.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
